apps/control-plane/pyproject.toml
[project]
name = "cadris-control-plane"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.116,<0.117",
  "httpx>=0.28,<0.29",
  "pydantic>=2.12,<3",
  "pydantic-settings>=2.7,<3",
  "python-multipart>=0.0.20,<0.0.23",
  "sqlalchemy>=2.0,<2.1",
  "uvicorn>=0.35,<0.36",
  "boto3>=1.35.0,<2",
  "google-auth>=2.0,<3",
  "openai>=2.26,<3",
  "stripe>=14.0,<15",
  "python-dotenv>=1.0,<2",
  "psycopg2-binary>=2.9,<3",
  "markdown>=3.7,<4",
  "bleach>=6.2,<7",
  "bcrypt>=4.2,<5",
  "python-pptx>=1.0,<2"
]

[tool.setuptools.packages.find]
include = ["cadris_cp*"]

[tool.uvicorn]
factory = false

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::pydantic._internal._generate_schema.UnsupportedFieldAttributeWarning",
]

[project.optional-dependencies]
test = [
  "pytest>=8.0.0",
  "pytest-asyncio>=0.23.0",
  "httpx>=0.28.1"
]
